Map Contribution
Contribution efforts featuring the map creation and webscraping efforts of Adam, James, and Nancy. HTML incorporation with other parts of project worked on by Adam, Todd, and Bryan.

Run app.py in inherit_withmap and navigate to the route /map to see the leaflet map in the main page template format.

Run map_app.py and navigate to route '/' to see the map, and '/scrape' to see the webscrape output.